+++ Runtime/SceneManager/SceneManager.cs
@@ -15,12 +15,13 @@ namespace ActionCode.SceneManagement
     public static class SceneManager
     {
         /// <summary>
-        /// Action fired when loading is started.
+        /// Action fired when loading is started. <see cref="LoadingScene"/> will be set.
         /// </summary>
         public static event Action OnLoadingStarted;
 
         /// <summary>
-        /// Action fired when loading is finished.
+        /// Action fired when loading is finished. 
+        /// <see cref="LoadingScene"/> will be null.
         /// </summary>
         public static event Action OnLoadingFinished;
 
@@ -36,12 +37,17 @@ namespace ActionCode.SceneManagement
         /// <summary>
         /// Whether a Scene loading process is happening. 
         /// </summary>
-        public static bool IsLoading { get; private set; }
+        public static bool IsLoading => LoadingScene != null;
 
         /// <summary>
         /// Whether the Scene loading process is locked. 
         /// </summary>
         public static bool IsLoadingLocked { get; private set; }
+
+        /// <summary>
+        /// The current Scene been loaded.
+        /// </summary>
+        public static Scene LoadingScene { get; private set; }
 
         /// <summary>
         /// Locks the Scene loading process.
@@ -62,27 +68,44 @@ namespace ActionCode.SceneManagement
         /// <param name="scene">The name or path of the Scene to load.</param>
         /// <param name="transition">The transition data to use.</param>
         public static void LoadScene(string scene, SceneTransition transition = null) =>
+            LoadScene(new Scene(scene), transition);
+
+        /// <summary>
+        /// <inheritdoc cref="LoadScene(string, SceneTransition)"/>
+        /// </summary>
+        /// <param name="scene">The Scene to load.</param>
+        /// <param name="transition"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='transition']"/></param>
+        public static void LoadScene(Scene scene, SceneTransition transition = null) =>
             _ = LoadSceneAsync(scene, transition);
 
         /// <summary>
         /// Loads the given Scene asynchronously using a Loading Scene and fade transitions.
         /// </summary>
-        /// <param name="scene">The name or path of the Scene to load</param>
-        /// <param name="transition">The transition data to use.</param>
+        /// <param name="scene"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='scene']"/></param>
+        /// <param name="transition"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='transition']"/></param>
         /// <returns>An asynchronously Task.</returns>
         public static async Task LoadSceneAsync(string scene, SceneTransition transition = null) =>
+            await LoadSceneAsync(new Scene(scene), transition);
+
+        /// <summary>
+        /// <inheritdoc cref="LoadSceneAsync(string, SceneTransition)"/>
+        /// </summary>
+        /// <param name="scene"><inheritdoc cref="LoadScene(Scene, SceneTransition)" path="/param[@name='scene']"/></param>
+        /// <param name="transition"><inheritdoc cref="LoadScene(string, SceneTransition)" path="/param[@name='transition']"/></param>
+        /// <returns><inheritdoc cref="LoadSceneAsync(string, SceneTransition)"/></returns>
+        public static async Task LoadSceneAsync(Scene scene, SceneTransition transition = null) =>
             await AwaitableCoroutine.Run(LoadSceneCoroutine(scene, transition));
 
-        private static IEnumerator LoadSceneCoroutine(string scene, SceneTransition transition)
+        private static IEnumerator LoadSceneCoroutine(Scene scene, SceneTransition transition)
         {
             if (IsLoading)
-                throw new Exception("Cannot load new Scene since other scene is being loaded.");
+                throw new Exception($"Cannot load {scene} since {LoadingScene} is being loaded.");
 
             if (transition == null) transition = ScriptableObject.CreateInstance<SceneTransition>();
 
             transition.Initialize();
 
-            IsLoading = true;
+            LoadingScene = scene;
             OnLoadingStarted?.Invoke();
 
             var hasLoadingScene = transition.HasLoadingScene();
@@ -98,7 +121,7 @@ namespace ActionCode.SceneManagement
                 // Check if LoadingScene was valid.
                 if (loadingSceneOperation == null)
                 {
-                    IsLoading = false;
+                    LoadingScene = null;
                     yield break;
                 }
 
@@ -110,12 +133,12 @@ namespace ActionCode.SceneManagement
 
             yield return new WaitForSeconds(transition.TimeBeforeLoading);
 
-            var loadingOperation = UnitySceneManager.LoadSceneAsync(scene);
+            var loadingOperation = UnitySceneManager.LoadSceneAsync(scene.path);
 
             // Check if scene was valid.
             if (loadingOperation == null)
             {
-                IsLoading = false;
+                LoadingScene = null;
                 yield break;
             }
 
@@ -137,7 +160,7 @@ namespace ActionCode.SceneManagement
 
             yield return transition.ScreenFader?.FadeIn();
 
-            IsLoading = false;
+            LoadingScene = null;
             OnLoadingFinished?.Invoke();
         }