+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- SceneManager.LoadScene functions using Scene
 - Scene class
 
 ## [4.0.0] - 2024-10-11